pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "omega-longread"
version = "0.1.0"
description = "Overlap-conditioned monotonic edit model for long-read correction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "pyyaml>=6.0",
  "numpy>=1.24",
  "pysam>=0.22",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
